+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -20,7 +20,7 @@ public class GenreService {
         return genreStorage.findAll();
     }
 
-    public Genre getById(int id) {
+    public Genre getById(Long id) {
         return genreStorage.getById(id);
     }
 }
